// frontend/src/app/Models/Horario/HorarioDetallePopupModel.ts
export class HorarioDetallePopupModel {
    HorarioDetalleId: number = 0;
    HorarioId: number = 0;
    Items: number = 0;
    HoraInicio: number = 0;
    MinutoInicio: number = 0;
    HoraTermino: number = 0;
    MinutoTermino: number = 0;
    TipoSecuenciaMarcacionId: number = 0;
    NomTipoSecuenciaMarcacion: String = '';
    FlaInicio: Boolean = false;
}